+++ lightly_studio_view/src/lib/components/Header/Header.svelte
@@ -153,8 +153,24 @@
 
 <style>
     :global {
-        @media screen and (max-width: 1300px) {
+        @media screen and (min-width: 1101px) {
+            .menu-trigger > .lucide-menu {
+                display: none;
+            }
+        }
+
+        @media screen and (max-width: 1100px) {
+            .menu-trigger {
+                width: 2.5rem;
+                justify-content: center;
+                padding: 0;
+            }
+
             .nav-button > span {
+                display: none;
+            }
+
+            .menu-trigger > .lucide-chevron-down {
                 display: none;
             }
         }